+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.39.4
 
-- Fix an issue where a link to a section was not opening at the correct section if an explicit preferred Zoom has been set in the PDF Pager module setting.s
+- Fix an issue where a link to a section was not opening at the correct section if an explicit preferred Zoom has been set in the PDF Pager module settings.
 
 ## 0.39.3